lib: Fix MEM_ALIGN to cope with huge allocations Attempting to allocate 2^32+1 bytes will look like it succeeds, as MEM_ALIGN will set alloc_size = 8. The caller will then think it's got 4 gig to play with.
e.g. t_malloc0 will wipe vast areas of memory before segfaulting, which might
include useful information we'd like in a corefile.

+++ src/lib/macros.h
@@ -19,7 +19,7 @@
 	(sizeof(arr) / sizeof((arr)[0]))
 
 #define MEM_ALIGN(size) \
-	(((size) + MEM_ALIGN_SIZE-1) & ~((unsigned int) MEM_ALIGN_SIZE-1))
+	(((size) + MEM_ALIGN_SIZE-1) & ~((size_t) MEM_ALIGN_SIZE-1))
 
 #define PTR_OFFSET(ptr, offset) \
 	((void *) (((unsigned char *) (ptr)) + (offset)))
